refactor: Extract data sources and list content composable
This commit refactors the data layer and UI components:

**Data Layer:**
- `AnimeRemoteDataSource.kt`: Extracted from `AnimeRepository.kt`. This class is now solely responsible for fetching anime data from the MyAnimeList (MAL) API. It handles network requests, authentication using SharedPreferences, and data transformation from API models to domain models.
- `AnimeLocalDataSource.kt`: Extracted from `AnimeRepository.kt`. This class is now responsible for interacting with the local Room database to store and retrieve anime data. It manages data access objects (DAO) and transforms data between database entities and domain models.
- `AnimeRepository.kt`: Updated to use the newly extracted `AnimeRemoteDataSource` and `AnimeLocalDataSource`. It now acts as a coordinator between these data sources, deciding whether to fetch data from the network or the local cache.

**UI Layer:**
- `AnimeListScreenContent.kt`: Introduced a new composable function that encapsulates the common UI logic for displaying an anime list. This includes handling the pull-to-refresh mechanism, loading states, and rendering the list of anime items. This component is now used by `MyListScreen.kt`.
- `MyListScreen.kt`: Simplified by delegating the list rendering and pull-to-refresh logic to the new `AnimeListScreenContent` composable. It now focuses on managing the ViewModel, state, and navigation.

**Other:**
- Added KDoc documentation to `MyListViewModel.kt`, `MyListScreen.kt`, `AnimeRemoteDataSource.kt`, `AnimeLocalDataSource.kt`, and `AnimeListScreenContent.kt` to improve code clarity and maintainability.
